+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - add link to putrecord-test repository
 - re-introduce proper WhiteWind handling
 - add WhiteWind tests
+- preserve existing properties/fields by default (e.g. title/visibility)
 
 ### v0.1.0